# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Portfolio Website API",
    "description": "API documentation for the Portfolio Website backend service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Post": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "content": { "type": "string" },
          "excerpt": { "type": "string" },
          "featured_image": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["draft", "published"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "tags": { 
            "type": "array",
            "items": { "type": "string" }
          },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "parent_id": { "type": "integer", "nullable": true }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/api/posts": {
      "get": {
        "summary": "Get all blog posts",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer" },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer" },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": { "type": "string", "enum": ["draft", "published"] },
            "description": "Filter by post status"
          },
          {
            "in": "query",
            "name": "category",
            "schema": { "type": "integer" },
            "description": "Filter by category ID"
          },
          {
            "in": "query",
            "name": "tag",
            "schema": { "type": "string" },
            "description": "Filter by tag name"
          }
        ],
        "responses": {
          "200": {
            "description": "List of blog posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Post" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new blog post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "excerpt": { "type": "string" },
                  "featured_image": { "type": "string", "format": "binary" },
                  "status": { "type": "string", "enum": ["draft", "published"] },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "category_ids": { "type": "array", "items": { "type": "integer" } }
                },
                "required": ["title", "content", "excerpt"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog post created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          }
        }
      }
    },
    "/api/posts/{slug}": {
      "get": {
        "summary": "Get a blog post by slug",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" },
            "description": "Post slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Blog post details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          },
          "404": {
            "description": "Post not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a blog post by slug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" },
            "description": "Post slug"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "content": { "type": "string" },
                  "excerpt": { "type": "string" },
                  "featured_image": { "type": "string", "format": "binary" },
                  "status": { "type": "string", "enum": ["draft", "published"] },
                  "tags": { "type": "array", "items": { "type": "string" } },
                  "category_ids": { "type": "array", "items": { "type": "integer" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog post updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a blog post by slug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" },
            "description": "Post slug"
          }
        ],
        "responses": {
          "204": {
            "description": "Blog post deleted"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "summary": "Get all categories",
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new category",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "parent_id": { "type": "integer" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },
    "/api/categories/{slug}": {
      "get": {
        "summary": "Get a category by slug",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" },
            "description": "Category slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a category by slug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" },
            "description": "Category slug"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "parent_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a category by slug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": { "type": "string" },
            "description": "Category slug"
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted"
          }
        }
      }
    }
  }
}